requirements.txt
requests>=2.19
python-dateutil>=2.7.3
mock; python_version < '3.3'
